+++ outputs.tf
@@ -9,7 +9,7 @@ output "atlantis_url_events" {
 }
 
 output "atlantis_allowed_repo_names" {
-  description = "Github repositories where webhook should be created"
+  description = "Git repositories where webhook should be created"
   value       = var.atlantis_allowed_repo_names
 }
 
@@ -30,10 +30,20 @@ output "webhook_secret" {
 
 output "alb_dns_name" {
   description = "Dns name of alb"
-  value       = module.alb.dns_name
+  value       = module.alb.this_lb_dns_name
+}
+
+output "alb_zone_id" {
+  description = "Zone ID of alb"
+  value       = module.alb.this_lb_zone_id
 }
 
 output "ecs_task_definition" {
   description = "Task definition for ECS service (used for external triggers)"
   value       = aws_ecs_service.atlantis.task_definition
+}
+
+output "ecs_security_group" {
+  description = "Security group assigned to ECS Service in network configuration"
+  value       = module.atlantis_sg.this_security_group_id
 }