+++ variables.tf
@@ -4,6 +4,12 @@ variable "name" {
   default     = "atlantis"
 }
 
+variable "internal" {
+  description = "Whether the load balancer is internal or external"
+  type        = bool
+  default     = false
+}
+
 variable "tags" {
   description = "A map of tags to use on all resources"
   type        = map(string)
@@ -14,6 +20,12 @@ variable "tags_sgs_additional" {
   description = "Additional tags to put on the security groups that are wide open"
   type        = map(string)
   default     = {}
+}
+
+variable "atlantis_fqdn" {
+  description = "FQDN of Atlantis to use. Set this only to override Route53 and ALB's DNS name."
+  type        = string
+  default     = null
 }
 
 # VPC
@@ -84,6 +96,48 @@ variable "alb_logging_enabled" {
   default     = false
 }
 
+variable "alb_authenticate_oidc" {
+  description = "Map of Authenticate OIDC parameters to protect ALB (eg, using Auth0). See https://www.terraform.io/docs/providers/aws/r/lb_listener.html#authenticate-oidc-action"
+  type        = any
+  default     = {}
+}
+
+variable "alb_authenticate_cognito" {
+  description = "Map of AWS Cognito authentication parameters to protect ALB (eg, using SAML). See https://www.terraform.io/docs/providers/aws/r/lb_listener.html#authenticate-cognito-action"
+  type        = any
+  default     = {}
+}
+
+variable "allow_unauthenticated_access" {
+  description = "Whether to create ALB listener rule to allow unauthenticated access for certain CIDR blocks (eg. allow GitHub webhooks to bypass OIDC authentication)"
+  type        = bool
+  default     = false
+}
+
+variable "allow_unauthenticated_access_priority" {
+  description = "ALB listener rule priority for allow unauthenticated access rule"
+  type        = number
+  default     = 10
+}
+
+variable "allow_github_webhooks" {
+  description = "Whether to allow access for GitHub webhooks"
+  type        = bool
+  default     = false
+}
+
+variable "github_webhooks_cidr_blocks" {
+  description = "List of CIDR blocks used by GitHub webhooks" # This is hardcoded to avoid dependency on github provider. Source: https://api.github.com/meta
+  type        = list(string)
+  default     = ["140.82.112.0/20", "185.199.108.0/22", "192.30.252.0/22"]
+}
+
+variable "whitelist_unauthenticated_cidr_blocks" {
+  description = "List of allowed CIDR blocks to bypass authentication"
+  type        = list(string)
+  default     = []
+}
+
 # ACM
 variable "certificate_arn" {
   description = "ARN of certificate issued by AWS ACM. If empty, a new ACM certificate will be created and validated using Route53 DNS"
@@ -102,6 +156,12 @@ variable "route53_zone_name" {
   description = "Route53 zone name to create ACM certificate in and main A-record, without trailing dot"
   type        = string
   default     = ""
+}
+
+variable "route53_record_name" {
+  description = "Name of Route53 record to create ACM certificate in and main A-record. If null is specified, var.name is used instead. Provide empty string to point root domain name to ALB."
+  type        = string
+  default     = null
 }
 
 variable "create_route53_record" {
@@ -143,7 +203,7 @@ variable "atlantis_bitbucket_user_token_ssm_parameter_name" {
 }
 
 variable "ssm_kms_key_arn" {
-  description = "ARN of KMS key to use for entryption and decryption of SSM Parameters. Required only if your key uses a custom KMS key and not the default key"
+  description = "ARN of KMS key to use for encryption and decryption of SSM Parameters. Required only if your key uses a custom KMS key and not the default key"
   type        = string
   default     = ""
 }
@@ -228,13 +288,25 @@ variable "atlantis_repo_whitelist" {
 }
 
 variable "atlantis_allowed_repo_names" {
-  description = "Github repositories where webhook should be created"
+  description = "Git repositories where webhook should be created"
   type        = list(string)
   default     = []
 }
 
 variable "allow_repo_config" {
   description = "When true allows the use of atlantis.yaml config files within the source repos."
+  type        = string
+  default     = "false"
+}
+
+variable "atlantis_log_level" {
+  description = "Log level that Atlantis will run with. Accepted values are: <debug|info|warn|error>"
+  type        = string
+  default     = "debug"
+}
+
+variable "atlantis_hide_prev_plan_comments" {
+  description = "Enables atlantis server --hide-prev-plan-comments hiding previous plan comments on update"
   type        = string
   default     = "false"
 }
@@ -292,24 +364,28 @@ variable "atlantis_bitbucket_base_url" {
 
 variable "custom_environment_secrets" {
   description = "List of additional secrets the container will use (list should contain maps with `name` and `valueFrom`)"
-  type        = list(map(string))
-  default     = []
+  type = list(object(
+    {
+      name      = string
+      valueFrom = string
+    }
+  ))
+  default = []
 }
 
 variable "custom_environment_variables" {
   description = "List of additional environment variables the container will use (list should contain maps with `name` and `value`)"
-  type        = list(map(string))
-  default     = []
+  type = list(object(
+    {
+      name  = string
+      value = string
+    }
+  ))
+  default = []
 }
 
 variable "security_group_ids" {
   description = "List of one or more security groups to be added to the load balancer"
   type        = list(string)
   default     = []
-}
-
-variable "aws_ssm_path" {
-  description = "AWS ARN prefix for SSM (public AWS region or Govcloud). Valid options: aws, aws-us-gov."
-  type        = string
-  default     = "aws"
 }